Action Cable for text channel message delete complete

+++ app/channels/chat_channel.rb
@@ -5,14 +5,10 @@ class ChatChannel < ApplicationCable::Channel
   end
 
   def speak(data)
-    # Creating a variable that is a Message string
-    # value of the string is the value of the body -> the value of message key in data
     message = TextChannelMessage.new(data['message'])
 
     if message.save
-      # socket converts the message string to an object
       socket = {message: message, type: 'message'}
-      # broadcasting this message object
       ChatChannel.broadcast_to('chat_channel', socket)
     end
   end
@@ -28,17 +24,17 @@ class ChatChannel < ApplicationCable::Channel
 
   def poof(data)
     message = TextChannelMessage.find_by(id: data['message']['id'])
-    if message.destroy
-      socket = {message: message, type: 'no_message'}
+
+    if message.author.id == data['currentUser']['id']
+      if message.destroy
+        socket = {message: message, type: 'no_message'}
+        ChatChannel.broadcast_to('chat_channel', socket)
+      end
+    else
+      socket = { type: 'inaction'}
       ChatChannel.broadcast_to('chat_channel', socket)
     end
   end
-
-  # def load
-  #   messages = TextChannelMessage.all.collect(&:body)
-  #   socket = { messages: messages, type: 'messages'}
-  #   ChatChannel.broadcast_to('chat_channel', socket)
-  # end
 
   def unsubscribed
     # Any cleanup needed when channel is unsubscribed
